Rework proto(s) target of Makefile
Change the "protos" target to "proto" since that is what I always want
to type.

Only run `go list -m` once instead of twice. Just a minor optimisation.

Ignore proto files starting with an underscore. This is a common thing
with the Go toolchain and is useful when the build does not require you
to specify filename explicitly. It means you can temporarily put files
out of the way by prefixing them with an underscore if they are not
ready to be built.