Make `list` command faster and sort output
Make the `list` command faster by not fetching the image size by default
as that requires one request per image and is quite slow for large
registries. Also make the default output be a form that can be cut and
paste when needing an image name, rather than using tabular output.
Tabular output is slower as it needs to wait for all output to be
generated before it produces output. Non-tabular form (the default)
produces output earlier as the repositories are iterated.

Also sort the output so it is easier to find what you're looking for.